fix(geo): address React Doctor and review findings
- drop manual useMemo in compiler-managed geo components
- move engineIconHtml into utils so engine-icon.tsx only exports components
- use toSorted for tooltip ranking (lib bumped to ES2023)
- raise the AI traffic pages fetch limit so path grouping sees every source row
