+++ apps/dashboard/src/components/geo/directions/direction-pages-table.tsx
@@ -1,7 +1,5 @@
 "use client";
 
-import { useMemo } from "react";
-
 import { EngineIcon } from "@/components/geo/engine-icon";
 import { Table, type TableColumn } from "@/components/motion/table";
 import { TruncateWithTooltip } from "@/components/truncate-with-tooltip";
@@ -15,46 +13,43 @@ import { formatDirectionCount } from "@/utils/geo-directions";
 import { tableHeightFor } from "@/utils/table";
 
 export function DirectionPagesTable({ className }: DirectionBlockProps) {
-  const columns = useMemo<TableColumn<GeoTrafficPage>[]>(
-    () => [
-      {
-        key: "path",
-        header: "Page",
-        width: "1.5fr",
-        sortable: true,
-        cell: (row) => (
-          <TruncateWithTooltip className="font-mono text-xs">
-            {row.path}
-          </TruncateWithTooltip>
-        ),
-      },
-      {
-        key: "source",
-        header: "Source",
-        width: "1fr",
-        sortable: true,
-        cell: (row) => (
-          <span className="flex min-w-0 items-center gap-2 text-sm">
-            <EngineIcon engine={row.source} />
-            <span className="truncate">{formatGeoSource(row.source)}</span>
-          </span>
-        ),
-        sortValue: (row) => formatGeoSource(row.source),
-      },
-      {
-        key: "visits",
-        header: "Visits",
-        width: "6.5rem",
-        sortable: true,
-        cell: (row) => (
-          <span className="text-sm tabular-nums">
-            {formatDirectionCount(row.visits)}
-          </span>
-        ),
-      },
-    ],
-    []
-  );
+  const columns: TableColumn<GeoTrafficPage>[] = [
+    {
+      key: "path",
+      header: "Page",
+      width: "1.5fr",
+      sortable: true,
+      cell: (row) => (
+        <TruncateWithTooltip className="font-mono text-xs">
+          {row.path}
+        </TruncateWithTooltip>
+      ),
+    },
+    {
+      key: "source",
+      header: "Source",
+      width: "1fr",
+      sortable: true,
+      cell: (row) => (
+        <span className="flex min-w-0 items-center gap-2 text-sm">
+          <EngineIcon engine={row.source} />
+          <span className="truncate">{formatGeoSource(row.source)}</span>
+        </span>
+      ),
+      sortValue: (row) => formatGeoSource(row.source),
+    },
+    {
+      key: "visits",
+      header: "Visits",
+      width: "6.5rem",
+      sortable: true,
+      cell: (row) => (
+        <span className="text-sm tabular-nums">
+          {formatDirectionCount(row.visits)}
+        </span>
+      ),
+    },
+  ];
 
   return (
     <div className={cn("flex flex-col gap-2", className)}>
