+++ apps/dashboard/src/components/geo/journeys-card.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { useMemo, useState } from "react";
+import { useState } from "react";
 
 import { EngineIcon } from "@/components/geo/engine-icon";
 import { JourneyDetailDialog } from "@/components/geo/journey-detail-dialog";
@@ -24,89 +24,84 @@ import { tableHeightFor } from "@/utils/table";
 export function JourneysCard({ journeys, organizationId }: JourneysCardProps) {
   const [selected, setSelected] = useState<GeoJourney | null>(null);
 
-  const columns = useMemo<TableColumn<GeoJourney>[]>(
-    () => [
-      {
-        key: "journeyId",
-        header: "Journey",
-        width: "7.5rem",
-        cell: (row) => (
-          <span
-            className="bg-muted text-muted-foreground rounded-sm px-1.5 py-0.5 font-mono text-xs"
-            title={row.journeyId}
-          >
-            {formatGeoJourneyChip(row.journeyId)}
-          </span>
-        ),
-      },
-      {
-        key: "source",
-        header: "Source",
-        width: "1fr",
-        sortable: true,
-        cell: (row) => (
-          <span className="flex min-w-0 items-center gap-2 text-sm">
-            <EngineIcon engine={row.source} />
-            <span className="truncate">{formatGeoSource(row.source)}</span>
-          </span>
-        ),
-        sortValue: (row) => formatGeoSource(row.source),
-      },
-      {
-        key: "pages",
-        header: "Pages",
-        width: "5.625rem",
-        sortable: true,
-        cell: (row) => (
-          <span className="text-sm tabular-nums">{row.pages}</span>
-        ),
-      },
-      {
-        key: "distinctPaths",
-        header: "Unique",
-        width: "5.625rem",
-        sortable: true,
-        cell: (row) => (
-          <span className="text-sm tabular-nums">{row.distinctPaths}</span>
-        ),
-      },
-      {
-        key: "span",
-        header: "Span",
-        width: "9.5rem",
-        cell: (row) => (
-          <span className="text-muted-foreground text-[0.6875rem] whitespace-nowrap tabular-nums">
-            {formatGeoJourneySpan(row.firstSeenAt, row.lastSeenAt)}
-          </span>
-        ),
-      },
-      {
-        key: "lastSeenAt",
-        header: "Last seen",
-        width: "9.375rem",
-        sortable: true,
-        cell: (row) => (
-          <span className="text-muted-foreground text-[0.6875rem] whitespace-nowrap tabular-nums">
-            {formatAiTrafficTimestamp(row.lastSeenAt)}
-          </span>
-        ),
-      },
-      {
-        key: "entryPath",
-        header: "Path",
-        width: "2fr",
-        cell: (row) => (
-          <JourneyPathTrail
-            className="flex-nowrap overflow-hidden"
-            limit={GEO_JOURNEY_TRAIL_TABLE_LIMIT}
-            paths={row.samplePaths}
-          />
-        ),
-        sortValue: (row) => row.samplePaths[0] ?? "",
-      },
-    ],
-    []
-  );
+  const columns: TableColumn<GeoJourney>[] = [
+    {
+      key: "journeyId",
+      header: "Journey",
+      width: "7.5rem",
+      cell: (row) => (
+        <span
+          className="bg-muted text-muted-foreground rounded-sm px-1.5 py-0.5 font-mono text-xs"
+          title={row.journeyId}
+        >
+          {formatGeoJourneyChip(row.journeyId)}
+        </span>
+      ),
+    },
+    {
+      key: "source",
+      header: "Source",
+      width: "1fr",
+      sortable: true,
+      cell: (row) => (
+        <span className="flex min-w-0 items-center gap-2 text-sm">
+          <EngineIcon engine={row.source} />
+          <span className="truncate">{formatGeoSource(row.source)}</span>
+        </span>
+      ),
+      sortValue: (row) => formatGeoSource(row.source),
+    },
+    {
+      key: "pages",
+      header: "Pages",
+      width: "5.625rem",
+      sortable: true,
+      cell: (row) => <span className="text-sm tabular-nums">{row.pages}</span>,
+    },
+    {
+      key: "distinctPaths",
+      header: "Unique",
+      width: "5.625rem",
+      sortable: true,
+      cell: (row) => (
+        <span className="text-sm tabular-nums">{row.distinctPaths}</span>
+      ),
+    },
+    {
+      key: "span",
+      header: "Span",
+      width: "9.5rem",
+      cell: (row) => (
+        <span className="text-muted-foreground text-[0.6875rem] whitespace-nowrap tabular-nums">
+          {formatGeoJourneySpan(row.firstSeenAt, row.lastSeenAt)}
+        </span>
+      ),
+    },
+    {
+      key: "lastSeenAt",
+      header: "Last seen",
+      width: "9.375rem",
+      sortable: true,
+      cell: (row) => (
+        <span className="text-muted-foreground text-[0.6875rem] whitespace-nowrap tabular-nums">
+          {formatAiTrafficTimestamp(row.lastSeenAt)}
+        </span>
+      ),
+    },
+    {
+      key: "entryPath",
+      header: "Path",
+      width: "2fr",
+      cell: (row) => (
+        <JourneyPathTrail
+          className="flex-nowrap overflow-hidden"
+          limit={GEO_JOURNEY_TRAIL_TABLE_LIMIT}
+          paths={row.samplePaths}
+        />
+      ),
+      sortValue: (row) => row.samplePaths[0] ?? "",
+    },
+  ];
 
   return (
     <InstrumentSection
