+++ apps/dashboard/src/constants/geo.ts
@@ -355,7 +355,7 @@ export const GEO_ANSWER_SYSTEM_PROMPT =
 export const AI_TRAFFIC_DEFAULT_DAYS = 30;
 export const AI_TRAFFIC_DEFAULT_LOG_LIMIT = 50;
 export const AI_TRAFFIC_DEFAULT_PAGES_LIMIT = 20;
-export const AI_TRAFFIC_PAGES_FETCH_LIMIT = 100;
+export const AI_TRAFFIC_PAGES_FETCH_LIMIT = 500;
 export const AI_TRAFFIC_LOG_FETCH_LIMIT = 200;
 export const GEO_TRAFFIC_SOURCES_PAGE_PARAM = "sourcesPage";
 export const GEO_TRAFFIC_MARKDOWN_COLUMN_KEY = "markdownVisits";
